tep-1002: clean ups and refactoring
Refactor (split) TacklerTxns to input handling
and ANTLR Parser Context handling parts.

Fix few warnings spotted by Scala 2.12.2.
